Change the variable type to return float

+++ src/Tugas2/IMatematika.java
@@ -5,6 +5,6 @@ public interface IMatematika {
     public void pertambahan(int a, int b);
     public void pengurangan(int a, int b);  
     public void perkalian(int a, int b);
-    public void pembagian(int a, int b);
+    public void pembagian(float a, float b);
 
 }
